feat: use an util exception handler instead of Nest.JS class
