+++ packages/twenty-server/src/graphql-config.service.ts
@@ -1,7 +1,6 @@
-import { Abstract, Injectable, Type } from '@nestjs/common';
+import { Injectable } from '@nestjs/common';
 import { ContextIdFactory, ModuleRef } from '@nestjs/core';
 import { GqlOptionsFactory } from '@nestjs/graphql';
-import { FILTER_CATCH_EXCEPTIONS } from '@nestjs/common/constants';
 
 import {
   YogaDriverConfig,
@@ -16,8 +15,8 @@ import { TokenService } from 'src/core/auth/services/token.service';
 import { CoreModule } from 'src/core/core.module';
 import { Workspace } from 'src/core/workspace/workspace.entity';
 import { WorkspaceFactory } from 'src/workspace/workspace.factory';
-import { exceptionFilters } from 'src/filters';
 import { ExceptionHandlerService } from 'src/integrations/exception-handler/exception-handler.service';
+import { globalExceptionHandler } from 'src/filters/utils/global-exception-handler.util';
 
 @Injectable()
 export class GraphQLConfigService
@@ -47,13 +46,24 @@ export class GraphQLConfigService
 
           return await this.createSchema(context, workspace);
         } catch (error) {
-          // Filter the exception
-          this.filterException(error);
+          if (error instanceof JsonWebTokenError) {
+            //mockedUserJWT
+            throw new GraphQLError('Unauthenticated', {
+              extensions: {
+                code: 'UNAUTHENTICATED',
+              },
+            });
+          }
 
-          // Send the unfiltered error to Sentry
-          this.exceptionHandlerService.captureException(error);
+          if (error instanceof TokenExpiredError) {
+            throw new GraphQLError('Unauthenticated', {
+              extensions: {
+                code: 'UNAUTHENTICATED',
+              },
+            });
+          }
 
-          throw error;
+          throw globalExceptionHandler(error, this.exceptionHandlerService);
         }
       },
       resolvers: { JSON: GraphQLJSON },
@@ -81,50 +91,5 @@ export class GraphQLConfigService
     );
 
     return await workspaceFactory.createGraphQLSchema(workspace.id);
-  }
-
-  private filterException(error: unknown) {
-    if (error instanceof JsonWebTokenError) {
-      //mockedUserJWT
-      throw new GraphQLError('Unauthenticated', {
-        extensions: {
-          code: 'UNAUTHENTICATED',
-        },
-      });
-    }
-
-    if (error instanceof TokenExpiredError) {
-      throw new GraphQLError('Unauthenticated', {
-        extensions: {
-          code: 'UNAUTHENTICATED',
-        },
-      });
-    }
-
-    /**
-     * Not really pretty but it works for now as we're outside of Nest.JS "flow" `APP_FILTER` is not triggered
-     * Loop through all the exception filters and check if the error is instance of any of the metadata
-     */
-    for (const exceptionFilter of exceptionFilters) {
-      // Extract the metadata from the exception filter
-      const metadata: Array<Type<any> | Abstract<any>> | undefined =
-        Reflect.getMetadata(FILTER_CATCH_EXCEPTIONS, exceptionFilter);
-
-      if (!metadata) {
-        continue;
-      }
-
-      // Loop through the metadata and check if the error is instance of any of the metadata
-      for (const err of metadata) {
-        // If the error is instance of the metadata, instantiate the exception filter and throw the error
-        if (error instanceof err) {
-          const filterInstance = new exceptionFilter(
-            this.exceptionHandlerService,
-          );
-
-          throw filterInstance.catchException(error as any);
-        }
-      }
-    }
   }
 }
